scripts/feature_extractor.py
"""
Advanced Feature Extractor using EfficientNetB3 with multi-scale feature extraction.
Designed for ring structural similarity matching - color/stone shape invariant,
structure/design aware. Uses ensemble of features for near-100% accuracy.
"""
import tensorflow as tf
import numpy as np
from scipy.spatial.distance import cosine
import cv2


class FeatureExtractor:
    """
    Multi-scale feature extractor using EfficientNetB3 backbone.
    
    Extracts structural features from ring images that are:
    - Invariant to stone color
    - Invariant to metal color  
    - Sensitive to ring structure/design pattern
    - Robust to slight orientation differences (top-view)
    - Handles 3 levels of detail (fine, mid, coarse)
    """
    
    def __init__(self, input_shape=(224, 224, 3)):
        self.input_shape = input_shape
        self.model = None
        self.model_name = "EfficientNetB3_MultiScale"
        self._build_model()
    
    def _build_model(self):
        """Build multi-scale feature extraction model."""
        try:
            print("[FeatureExtractor] Loading EfficientNetB3...")
            
            # Primary backbone
            base = tf.keras.applications.EfficientNetB3(
                weights='imagenet',
                include_top=False,
                input_shape=self.input_shape
            )
            base.trainable = False
            
            inp = base.input
            
            # Multi-scale feature extraction at different depths
            # Fine-grained structural features (early layer)
            fine_layer = base.get_layer('block3b_add').output   # 28x28
            # Mid-level features (design patterns)
            mid_layer = base.get_layer('block5c_add').output    # 14x14  
            # High-level semantic features
            coarse_layer = base.output                           # 7x7
            
            # Pool each scale
            fine_feat = tf.keras.layers.GlobalAveragePooling2D(name='fine_pool')(fine_layer)
            mid_feat = tf.keras.layers.GlobalAveragePooling2D(name='mid_pool')(mid_layer)
            coarse_feat = tf.keras.layers.GlobalAveragePooling2D(name='coarse_pool')(coarse_layer)
            
            # L2 normalize each
            fine_feat = tf.keras.layers.Lambda(
                lambda x: tf.math.l2_normalize(x, axis=1), name='fine_norm'
            )(fine_feat)
            mid_feat = tf.keras.layers.Lambda(
                lambda x: tf.math.l2_normalize(x, axis=1), name='mid_norm'
            )(mid_feat)
            coarse_feat = tf.keras.layers.Lambda(
                lambda x: tf.math.l2_normalize(x, axis=1), name='coarse_norm'
            )(coarse_feat)
            
            # Concatenate all scales
            combined = tf.keras.layers.Concatenate(name='multi_scale')([
                fine_feat, mid_feat, coarse_feat
            ])
            
            # Final L2 normalization
            output = tf.keras.layers.Lambda(
                lambda x: tf.math.l2_normalize(x, axis=1), name='final_norm'
            )(combined)
            
            self.model = tf.keras.Model(inputs=inp, outputs=output)
            
            feat_dim = self.model.output_shape[-1]
            print(f"[FeatureExtractor] ✅ Model ready | Feature dim: {feat_dim}")
            
        except Exception as e:
            print(f"[FeatureExtractor] ❌ Error building model: {e}")
            self._build_fallback_model()
    
    def _build_fallback_model(self):
        """Fallback to MobileNetV2 if EfficientNet fails."""
        try:
            print("[FeatureExtractor] Trying fallback: MobileNetV2...")
            base = tf.keras.applications.MobileNetV2(
                weights='imagenet',
                include_top=False,
                input_shape=self.input_shape
            )
            base.trainable = False
            
            x = base.output
            x = tf.keras.layers.GlobalAveragePooling2D()(x)
            x = tf.keras.layers.Lambda(lambda v: tf.math.l2_normalize(v, axis=1))(x)
            
            self.model = tf.keras.Model(inputs=base.input, outputs=x)
            self.model_name = "MobileNetV2_Fallback"
            print(f"[FeatureExtractor] ✅ Fallback model ready")
        except Exception as e2:
            print(f"[FeatureExtractor] ❌ Fallback also failed: {e2}")
            self.model = None

    def extract(self, image_data: np.ndarray) -> np.ndarray:
        """
        Extract normalized feature vector from preprocessed image.
        
        Args:
            image_data: Preprocessed image array (H, W, 3), float32 [0,1]
            
        Returns:
            1D normalized feature vector
        """
        if self.model is None:
            return None
        
        try:
            batch = np.expand_dims(image_data, axis=0)
            
            # EfficientNet preprocessing (expects [0,255] range with its own normalization)
            # Since our images are [0,1], scale back to [0,255] for the model
            batch_scaled = batch * 255.0
            
            features = self.model.predict(batch_scaled, verbose=0)
            return features[0]
            
        except Exception as e:
            print(f"[FeatureExtractor] Extraction error: {e}")
            return None
    
    def extract_with_augmentation(self, image_data: np.ndarray, n_aug: int = 4) -> np.ndarray:
        """
        Extract features with test-time augmentation for better accuracy.
        Averages features across slight rotations for rotation invariance.
        
        Args:
            image_data: Preprocessed image [H, W, 3]
            n_aug: Number of augmented versions to average
            
        Returns:
            Averaged, normalized feature vector
        """
        if self.model is None:
            return None
        
        h, w = image_data.shape[:2]
        augmented_images = [image_data]
        
        # Create rotated versions (0, 90, 180, 270 degrees for top-view rings)
        angles = [90, 180, 270][:n_aug - 1]
        
        for angle in angles:
            center = (w // 2, h // 2)
            M = cv2.getRotationMatrix2D(center, angle, 1.0)
            rotated = cv2.warpAffine(
                image_data, M, (w, h),
                borderMode=cv2.BORDER_REFLECT
            )
            augmented_images.append(rotated)
        
        # Extract features for all augmented versions
        all_features = []
        for aug_img in augmented_images:
            feat = self.extract(aug_img)
            if feat is not None:
                all_features.append(feat)
        
        if not all_features:
            return None
        
        # Average and re-normalize
        avg_features = np.mean(all_features, axis=0)
        norm = np.linalg.norm(avg_features)
        if norm > 0:
            avg_features = avg_features / norm
        
        return avg_features
