requirements.txt
streamlit
tensorflow-cpu
opencv-python
numpy
Pillow
scikit-learn
scipy
tqdm
ml-dtypes
